Optimize PATCH operations to avoid scan
The SPciO key is split into SPci (key) and iO (value).
This allows for deterministically calculating the triples
that are to be removed from a graph. By knowing the key
of the triples that are to be removed, a scan is avoided.

Keys are stored as follows. Breaking change

name.S subject predicate index
name.P predicate isId object index subject
name.O isId object subject predicate index